Complete update of all my work today
Finished Project. Copied starter code in and created the needed questions and implemented the needed code located in the generateMarkdown.js file and index.js file. I struggled to get node index.js to work but as soon as I realized that the generateMarkdown.js and index.js files had to both be in my util folder it worked without isses