Updated example READ.me file
As stated in title I updated my project to have an example READ.me and also moved the files around to make it easier for the client to navigate my app. I deleted my "develop" folder and moved everything previously inside out into the root of the directory. I also moved the "index.js" file out of the utils folder and into the root directory as well. Hopefully these adjustments will be enough to get 100%.